# api-docs/openAPI-specification.yml
openapi: 3.0.1
info:
  version: "1.0.0"
  title: RUMbus API
  description: >-
    RUMbus is a system to provide real-time information on the location of
    buses, as well as estimated arrival times at stops for the transportation
    service of the University of Puerto Rico at Mayaguez (UPRM).
...
paths:
...
  /trips/{trip_id}/events:
    post:
      tags:
        - Trips
      summary: Create trip event
      parameters:
        - in: path
          name: trip_id
          schema:
            type: string
          required: true
          description: ID of the trip
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [description, type]
              properties:
                description:
                  type: string
                type:
                  type: string
            examples:
              'Body':
                value: "{ \r\n    \"description\": \"test\",\r\n    \"type\": \"accidente\"\r\n}\r\n"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                'Created':
                  value: '{"message":"Trip event created!"}'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                'Missing param':
                  value: >-
                    {"code":"BadRequest","message":"Missing required param:
                    type"}
                'Bad Request':
                  value: '{"code":"BadRequest","message":"Need to provide a trip ID."}'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                'Not Found':
                  value: >-
                    {"code":"NotFound","message":"Trip with id:
                    4de277ac-f535-46f2-8773-4202a58cb4q1 not found"}
        '500':
          $ref: '#/components/responses/500'
    get:
      tags:
        - Trips
      summary: Get trip event
      parameters:
        - in: path
          name: trip_id
          schema:
            type: string
          required: true
          description: ID of the trip
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    id:
                      type: string
                    trip_id:
                      type: string
                    description:
                      type: string
                    type:
                      type: string
                    updated_on:
                      type: string
                    created_on:
                      type: string
              examples:
                'OK':
                  value: >-
                    [{"id":"871d8c53-1902-46ca-b9ce-7487b0b579da","trip_id":"4de277ac-f535-46f2-8773-4202a58cb461","description":"test","type":"accidente","updated_on":"2020-04-06T04:05:14.000Z","created_on":"2020-04-06T04:05:14.000Z"}]
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                'Bad Request':
                  value: '{"code":"BadRequest","message":"Need to provide a trip ID."}'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                'Not Found':
                  value: >-
                    {"code":"NotFound","message":"Trip with id:
                    4de277ac-f535-46f2-8773-4202a58cb46a not found"}
        '500':
          $ref: '#/components/responses/500'
...
  /routes/search:
    get:
      tags:
        - Routes
      description: Filter routes specifying any combination of query parameters.
      summary: Search routes with the specified parameters
      parameters:
        - in: query
          name: id
          schema:
            type: string
          description: Route id.
        - in: query
          name: name
          schema:
            type: string
          description: Route name.
        - in: query
          name: description
          schema:
            type: string
          description: Route description.
        - in: query
          name: geojson
          schema:
            type: boolean
          example: 'true'
          description: Change the default geometry (polyline) to geopoints of all the returned routes.
        - in: query
          name: stops
          schema:
            type: string
          description: Filter routes that include the specified stop(s). Stop coordinates are separated by a comma (,) where the longitude is specified first and then the latitude (i.e. "/search?stops=[lon,lat]"). You can specify multiple stops by separating them by semicolon (;) (i.e. "/search?stops=[lon,lat];[lon,lat]..."). (e.g. "/search?stops=-67.133728,18.215803", "/search?stops=-67.133728,18.215803;-67.1396338,18.2108147...")
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    description:
                      type: string
                    geometry:
                      type: string
                    stops:
                      type: array
                      items:
                        type: object
                        properties:
                          lon:
                            type: number
                          lat:
                            type: number
              examples:
                '/routes/search?name=Route 1':
                  value: >-
                    [{"id":"b7cdc823-7106-469d-8ae3-41ed2f259f33","name":"Route 1","description":"This is a testing route.","updated_on":"2020-03-23T19:27:11.000Z","created_on":"2020-03-23T19:27:11.000Z","stops":[{"lon":-67.133728,"lat":18.215803},{"lon":-67.1396338,"lat":18.2108147}],"geometry":"oxdnBl`wxKS^~@lDh@~@ZNtCh@jAl@^h@RbBPf@j@p@vChAXR`AfCNTxD`Bj@q@~CrA{AdEzAeE_DsAk@p@yDaBOUaAgCYSwCiAk@q@Qg@ScB_@i@kAm@uCi@[Oi@_A_AmDR_@"}]
                '/routes/search?name=Route 1&geojson=true':
                  value: >-
                    [{"id":"b7cdc823-7106-469d-8ae3-41ed2f259f33","name":"Route 1","description":"This is a testing route.","updated_on":"2020-03-23T19:27:11.000Z","created_on":"2020-03-23T19:27:11.000Z","stops":[{"lon":-67.133728,"lat":18.215803},{"lon":-67.1396338,"lat":18.2108147}],"geometry":[[-67.133672,18.215916],[-67.133831,18.216019],[-67.134699,18.215696],[-67.135021,18.215492],[-67.135098,18.215348],[-67.135313,18.214599],[-67.135539,18.21422],[-67.13575,18.214058],[-67.136249,18.213964],[-67.136446,18.213868],[-67.136698,18.213646],[-67.137072,18.212885],[-67.137172,18.212762],[-67.137849,18.212433],[-67.137958,18.212346],[-67.13845,18.21142],[-67.1382,18.211199],[-67.138615,18.210396],[-67.139613,18.210856],[-67.138615,18.210396],[-67.1382,18.211199],[-67.13845,18.21142],[-67.137958,18.212346],[-67.137849,18.212433],[-67.137172,18.212762],[-67.137072,18.212885],[-67.136698,18.213646],[-67.136446,18.213868],[-67.136249,18.213964],[-67.13575,18.214058],[-67.135539,18.21422],[-67.135313,18.214599],[-67.135098,18.215348],[-67.135021,18.215492],[-67.134699,18.215696],[-67.133831,18.216019],[-67.133672,18.215916]]}]
        '500':
          $ref: '#/components/responses/500'
  /routes/{route_id}:
    get:
      tags:
        - Routes
      parameters:
        - in: path
          name: route_id
          schema:
            type: string
          required: true
          description: ID of the route to fetch.
        - name: geojson
          in: query
          schema:
            type: boolean
          example: 'true'
          description: Change the default geometry (polyline) to geopoints of the returned route.
      summary: Get route
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                  updated_on:
                    type: string
                  created_on:
                    type: string
                  geometry:
                    type: string
                  stops:
                    type: array
                    items:
                      type: object
                      properties:
                        lon:
                          type: number
                        lat:
                          type: number
              examples:
                '/routes/{route_id}':
                  value: >-
                    {"id":"b7cdc823-7106-469d-8ae3-41ed2f259f33","name":"Route 1","description":"This is a testing route.","updated_on":"2020-03-23T19:27:11.000Z","created_on":"2020-03-23T19:27:11.000Z","stops":[{"lon":-67.133728,"lat":18.215803},{"lon":-67.1396338,"lat":18.2108147}],"geometry":"oxdnBl`wxKS^~@lDh@~@ZNtCh@jAl@^h@RbBPf@j@p@vChAXR`AfCNTxD`Bj@q@~CrA{AdEzAeE_DsAk@p@yDaBOUaAgCYSwCiAk@q@Qg@ScB_@i@kAm@uCi@[Oi@_A_AmDR_@"}
                '/routes/{route_id}?geojson=true':
                  value: >-
                    {"id":"b7cdc823-7106-469d-8ae3-41ed2f259f33","name":"Route 1","description":"This is a testing route.","updated_on":"2020-03-23T19:27:11.000Z","created_on":"2020-03-23T19:27:11.000Z","stops":[{"lon":-67.133728,"lat":18.215803},{"lon":-67.1396338,"lat":18.2108147}],"geometry":[[-67.133672,18.215916],[-67.133831,18.216019],[-67.134699,18.215696],[-67.135021,18.215492],[-67.135098,18.215348],[-67.135313,18.214599],[-67.135539,18.21422],[-67.13575,18.214058],[-67.136249,18.213964],[-67.136446,18.213868],[-67.136698,18.213646],[-67.137072,18.212885],[-67.137172,18.212762],[-67.137849,18.212433],[-67.137958,18.212346],[-67.13845,18.21142],[-67.1382,18.211199],[-67.138615,18.210396],[-67.139613,18.210856],[-67.138615,18.210396],[-67.1382,18.211199],[-67.13845,18.21142],[-67.137958,18.212346],[-67.137849,18.212433],[-67.137172,18.212762],[-67.137072,18.212885],[-67.136698,18.213646],[-67.136446,18.213868],[-67.136249,18.213964],[-67.13575,18.214058],[-67.135539,18.21422],[-67.135313,18.214599],[-67.135098,18.215348],[-67.135021,18.215492],[-67.134699,18.215696],[-67.133831,18.216019],[-67.133672,18.215916]]}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '/routes/':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '/routes/{route_id}':
                  value: >-
                    {"code":"NotFound","message":"TripRoute with id
                    a1b2c3d4-e5f6g7h8 not found."}
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Routes
      summary: Update route by its id
      parameters:
        - in: path
          name: route_id
          schema:
            type: string
          required: true
          description: ID of the route to update
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                stops:
                  type: array
                  items:
                    type: object
                    properties:
                      lon:
                        type: number
                      lat:
                        type: number
                      name:
                        type: string
            examples:
              '0':
                value: "{\n    \"name\": \"Other name\",\t\n    \"stops\": [\n        {\n            \"lon\": -67.141934,\n            \"lat\": 18.211754\n        },\n        {\n            \"lon\": -67.139084,\n            \"lat\": 18.209623\n        }\n    ]\n}"
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '/routes/':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Routes
      summary: Delete route by its id
      parameters:
        - in: path
          name: route_id
          schema:
            type: string
          required: true
          description: ID of the route to delete
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '/routes/':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '500':
          $ref: '#/components/responses/500'
...
components:
  responses:
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
          examples:
            'Server Error':
              value: '{"code":"InternalServer","message":"An internal server error ocurred"}'